chore(rq): make rq internal
