+++ releasenotes/notes/make-rq-imp-details-internal-c54afd12ba7b4bb7.yaml
@@ -1,0 +1,4 @@
+---
+deprecations:
+  - |
+    rq: Ensures the implementation details of the rq integration are internal to ddtrace. This integration should only be enabled and configured via ``ddtrace.patch(..)``, ``import ddtrace.auto`` or the ``ddtrace-run`` command.
